+++ test/__scenarios__/z-number.input.ts
@@ -7,3 +7,6 @@ const integerNumber = z.number().int();
 const integerNumberWithAttrs = z.number().max(23).int().gt(10);
 
 type Integer = z.infer<typeof integerNumber>;
+
+// Coerce should be untouched, it doesn't have .int()
+const coercedInteger = z.coerce.number().positive().int();