+++ test/__scenarios__/z-number.output.ts
@@ -7,3 +7,6 @@ const integerNumber = z.int();
 const integerNumberWithAttrs = z.int().max(23).gt(10);
 
 type Integer = z.infer<typeof integerNumber>;
+
+// Coerce should be untouched, it doesn't have .int()
+const coercedInteger = z.coerce.number().positive().int();